test: dev 리네임에 맞춰 승인 대표이미지 테스트를 UpdateWorkspaceRequestStatusTests로 통합
ApproveWorkspaceRequest→UpdateWorkspaceRequestStatus 리네임에 따라
옛 ApproveWorkspaceRequestTests의 대표이미지 마이그레이션 케이스를
UpdateWorkspaceRequestStatusTests로 이전하고, ACTIVATED 케이스에
누락된 stub을 보강.
